move clean_cache util method
